requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
networkx==3.3
google-generativeai==0.7.2
python-dotenv==1.0.1
pypdf==4.2.0
